gh-3055 Rename classes
