+++ stroom-cache/stroom-cache-api/src/main/java/stroom/cache/api/LoadingStroomCache.java
