+++ stroom-cache/stroom-cache-impl/src/main/java/stroom/cache/impl/AbstractStroomCache.java
